requirements/base.txt
dj-database-url==0.5.0
Django==3.2.4
psycopg2-binary
python-decouple==3.4
pytz==2021.1
djangorestframework-simplejwt
PyJWT
djangorestframework
djoser
